Refactor: 새 파일명 생성 함수 분리 재사용을 위하여 새로운 파일명 생성 함수를 파일 분리
